requirements.txt
langchain
langchain-core
langchain_community
langchain_openai
langgraph
openai
python-dotenv
rich
ipython
streamlit

